Consolidate data handling util routines
