+++ tests/test_tiler.py
@@ -9,7 +9,7 @@ from attr import dataclass
 import numpy as np
 import rioxarray
 from deadtrees.deployment.tiler import divisible_without_remainder, inspect_tile, Tiler
-from deadtrees.utils.tiletransforms import (
+from deadtrees.utils.data_handling import (
     make_blocks_vectorized,
     unmake_blocks_vectorized,
 )
